Added data file, updated ex1.f90 to read it
A new data file containing time variables and initial
conditions for 3 particles has been added.

My ex1.f90 has been modified in consequence to include
the necessary statements to open, read and close this
file. Print statements are added to make sure the data
is being read correctly (it seems so). Not yet modified
to include the saving to a file instead of printing.