Added output file writing
The program ex1.f90 now also writes the results
to a .txt file, and I have checked that the
results are properly written from the calculated
data.

I might regret the format when I try to plot
the data with Python, so it might still be
subjected to changes.

Apropos of that, I still have not checked
whether the results actually make sense. Will
do so by plotting the data with Python.