Changed output format and python plots
Changed the format of the writing to file output
so that the table is more easily read as a space
separated value file. First column is time, then
each particle x,y,z in order of particles.

Created a very simple Python notebook (might
change to .py later) to visualize the time
evolution of the position of the particles.
Needs more work to optimize and automatize
the plotting for varying numbers of particles.