Working version of ex2 (unparallelized)
New additions and corrections. The code is now working
and works as expected for the stable three-body example.

Added a new Python script to generate initial conditions
(which may be changed into a Fortran script later on)
and tested the code for N=20.

Modified the plot_nbody Python script to animate any
number of particles without having to write a line
for every particle (generalized to a for loop).